requirements.txt
Django==5.2.1
djangorestframework==3.16.0
python-dotenv==1.1.0
psycopg2==2.9.10
requests==2.32.3
Pillow==11.2.1
django-filter==25.1
pytest-django==4.11.1
pytest-factoryboy==2.7.0